+++ Core/AssemblyInfo.cs
@@ -1,7 +1,7 @@
 using System.Runtime.CompilerServices;
 using Unity.Cloud.Common;
 
-[assembly: ApiSourceVersion("com.unity.cloud.identity", "1.1.0")]
+[assembly: ApiSourceVersion("com.unity.cloud.identity", "1.2.0-exp.1")]
 #if !(UC_NUGET)
 [assembly: InternalsVisibleTo("Unity.Cloud.Identity.Tests")]
 [assembly: InternalsVisibleTo("Unity.Cloud.Identity.Editor")]