+++ Core/UserInformation/ICloudStorageJsonProvider.cs
@@ -1,0 +1,13 @@
+using System.Threading;
+using System.Threading.Tasks;
+using Unity.Cloud.Common;
+
+namespace Unity.Cloud.Identity
+{
+    internal interface ICloudStorageJsonProvider
+    {
+        public Task<CloudStorageUsageJson> GetCloudStorageUsageAsync(CancellationToken cancellationToken);
+
+        public Task<CloudStorageEntitlementsJson> GetCloudStorageEntitlementsAsync(CancellationToken cancellationToken);
+    }
+}