+++ Core/UserInformation/Models/CloudStorageUsage.cs
@@ -1,0 +1,15 @@
+namespace Unity.Cloud.Identity
+{
+    public class CloudStorageUsage : ICloudStorageUsage
+    {
+        public ulong UsageBytes { get; set; }
+        public ulong TotalStorageQuotaBytes { get; set; }
+
+        internal CloudStorageUsage(CloudStorageUsageJson cloudStorageUsageJson)
+        {
+            UsageBytes = cloudStorageUsageJson.UsageBytes;
+            TotalStorageQuotaBytes = cloudStorageUsageJson.TotalStorageQuotaBytes;
+        }
+
+    }
+}