+++ Core/UserInformation/Models/CloudStorageUsageJson.cs
@@ -1,0 +1,9 @@
+namespace Unity.Cloud.Identity
+{
+    internal class CloudStorageUsageJson
+    {
+        public ulong UsageBytes { get; set; }
+
+        public ulong TotalStorageQuotaBytes { get; set; }
+    }
+}