+++ Core/UserInformation/Models/ICloudStorageUsage.cs
@@ -1,0 +1,9 @@
+namespace Unity.Cloud.Identity
+{
+    public interface ICloudStorageUsage
+    {
+        public ulong UsageBytes { get; set; }
+
+        public ulong TotalStorageQuotaBytes { get; set; }
+    }
+}