+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0-exp.1] - 2024-04-19
+
+### Added
+- [Experimental] `ICloudStorageInfoProvider`, `ICloudStorageUsage` `ICloudStorageEntitlements` and `ICloudStorageEntitlement` interfaces.
+- [Experimental] `CloudStorageUsage`, `CloudStorageEntitlements` and `CloudStorageEntitlement` classes.
+- [Experimental] Add new `UnityEditorCloudServiceAuthorizer` supporting domain reload and transient disconnection in the editor.
+
+### Changed
+- [Experimental] `IOrganization` inherits `ICloudStorageInfoProvider`.
+
 ## [1.1.0] - 2024-04-05
 
 ### Added